Crypto share bug fixed
1. User input verification for Crypto html pages;
2. Allow FloatNumber for crypto share

+++ aerc_project/aerc_website/views.py
@@ -1098,9 +1098,9 @@ def crypto(request):
             purchase_date = request.POST.get('purchase_date', "")
             crypto = Crypto.objects.get(id=id)
             if buy_or_sell == "sell":
-                share = -int(request.POST.get('share', 0))
+                share = -float(request.POST.get('share', 0))
             else:
-                share = int(request.POST.get('share', 0))
+                share = float(request.POST.get('share', 0))
             transaction = CryptoTransaction(
                 crypto=crypto,
                 share=share,
